Tweaked start sound timing for the receiver
Original position of play_sound for the receiver meant that the OS seemed to be muting the sound since facetime was active. Moving the play sound to the start of the portal_open time means that it'll consistently play an audible alert to the receiver so that people know the portal is going to be opening shortly.